cleanup mkdocstring plugin
