+++ allowed_errors.json
@@ -2,7 +2,6 @@
   "allowedFlake8Rules": {
     "WPS212": 1,
     "WPS211": 1,
-    "WPS232": 1,
     "WPS361": 3,
     "WPS231": 2,
     "WPS421": 2,
