+++ packages/python/m/cli/args.py
@@ -63,8 +63,9 @@ def Arg(  # noqa: N802, WPS211
         validator: Function to validate the argument.
 
     Returns:
-        A new [`FieldInfo`][pydantic.fields.FieldInfo], the return annotation is
-        `Any` so `Arg` can be used on type annotated fields without causing a typing error.
+        A new [`FieldInfo`][pydantic.fields.FieldInfo]. The return annotation is
+            `Any` so `Arg` can be used on type annotated fields without causing
+            a typing error.
     """
     extras = {
         'positional': positional,
@@ -80,7 +81,7 @@ def Arg(  # noqa: N802, WPS211
     )
 
 
-def ArgProxy(*args, **kwargs) -> Any:  # noqa: N802
+def ArgProxy(*args: Any, **kwargs: Any) -> Any:  # noqa: N802
     """Wrap function to provide all argparse inputs.
 
     This is a escape hatch and does not provide any typings.
@@ -91,8 +92,9 @@ def ArgProxy(*args, **kwargs) -> Any:  # noqa: N802
         kwargs: The keyword arguments to argparse add arguments.
 
     Returns:
-        A new [`FieldInfo`][pydantic.fields.FieldInfo], the return annotation is
-        `Any` so `Arg` can be used on type annotated fields without causing a typing error.
+        A new [`FieldInfo`][pydantic.fields.FieldInfo]. The return annotation is
+            `Any` so `Arg` can be used on type annotated fields without causing
+            a typing error.
     """
     return FieldInfo.from_field(
         json_schema_extra={
